implement JoypadModule
