+++ src/JoypadModule/app/dcmWalkingJoypad/joypad.ini
@@ -1,0 +1,33 @@
+name            joypad
+period          0.05
+
+# Polydriver options
+# remote depends on the name of the joypad (xbox ps4 ...)
+
+# device    JoypadControlClient
+# local     /joypadInput
+# remote    /joypadDevice/xbox
+
+device    SDLJoypad
+sticks    0
+
+
+# Joypad options
+deadzone        0.05
+fullscale       1.0
+
+# scale factors
+scale_x         5.0
+scale_y         5.0
+
+# Port names
+rpcServerPort_name              /walking-coordinator/rpc
+rpcClientPort_name              /rpc
+
+# The goal is the desired position of the robot CoM. The desired position is a
+# vector containing the x and the y coordinate expressed in a frame rigidly
+# attached to the robot.
+# The x-axis points forward
+# The y-axis points on the left
+robotGoalInputPort_name         /walking-coordinator/goal:i
+robotGoalOutputPort_name        /goal:o
